Refactor containers code into a workloads package and separate the config package from the operator package
