--- pkg/crds/controllers/init.go
@@ -1,31 +1,0 @@
-/*
-Copyright 2021 Cortex Labs, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package controllers
-
-import "github.com/cortexlabs/cortex/pkg/types/clusterconfig"
-
-// Useful global variables
-var (
-	OperatorMetadata *clusterconfig.OperatorMetadata
-	ClusterConfig    *clusterconfig.Config
-)
-
-// Init initializes the package level global variables
-func Init(clusterConfig *clusterconfig.Config, operatorMetadata *clusterconfig.OperatorMetadata) {
-	OperatorMetadata = operatorMetadata
-	ClusterConfig = clusterConfig
-}
