+++ pkg/crds/main.go
@@ -22,9 +22,9 @@ import (
 	"os"
 	"strings"
 
+	"github.com/cortexlabs/cortex/pkg/config"
 	"github.com/cortexlabs/cortex/pkg/consts"
 	batch "github.com/cortexlabs/cortex/pkg/crds/apis/batch/v1alpha1"
-	"github.com/cortexlabs/cortex/pkg/crds/controllers"
 	batchcontrollers "github.com/cortexlabs/cortex/pkg/crds/controllers/batch"
 	awslib "github.com/cortexlabs/cortex/pkg/lib/aws"
 	"github.com/cortexlabs/cortex/pkg/lib/hash"
@@ -131,13 +131,13 @@ func main() {
 	}
 
 	// initialize some of the global values for the k8s helpers
-	controllers.Init(clusterConfig, operatorMetadata)
+	config.InitConfigs(clusterConfig, operatorMetadata)
 
-	config := batchcontrollers.BatchJobReconcilerConfig{}.ApplyDefaults()
+	reconcilerConfig := batchcontrollers.BatchJobReconcilerConfig{}.ApplyDefaults()
 
 	if err = (&batchcontrollers.BatchJobReconciler{
 		Client:        mgr.GetClient(),
-		Config:        config,
+		Config:        reconcilerConfig,
 		Log:           ctrl.Log.WithName("controllers").WithName("BatchJob"),
 		ClusterConfig: clusterConfig,
 		AWS:           awsClient,
