+++ pkg/operator/resources/job/batchapi/s3_iterator.go
@@ -18,9 +18,9 @@ package batchapi
 
 import (
 	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/cortexlabs/cortex/pkg/config"
 	"github.com/cortexlabs/cortex/pkg/lib/aws"
 	"github.com/cortexlabs/cortex/pkg/lib/errors"
-	"github.com/cortexlabs/cortex/pkg/operator/config"
 	"github.com/cortexlabs/cortex/pkg/operator/schema"
 	"github.com/gobwas/glob"
 )
