+++ pkg/operator/resources/job/cache.go
@@ -20,7 +20,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/cortexlabs/cortex/pkg/operator/config"
+	"github.com/cortexlabs/cortex/pkg/config"
 	"github.com/cortexlabs/cortex/pkg/types/spec"
 	"github.com/cortexlabs/cortex/pkg/types/userconfig"
 )
