+++ pkg/operator/resources/job/taskapi/job_status.go
@@ -17,7 +17,7 @@ limitations under the License.
 package taskapi
 
 import (
-	"github.com/cortexlabs/cortex/pkg/operator/config"
+	"github.com/cortexlabs/cortex/pkg/config"
 	"github.com/cortexlabs/cortex/pkg/operator/operator"
 	"github.com/cortexlabs/cortex/pkg/operator/resources/job"
 	"github.com/cortexlabs/cortex/pkg/types/spec"
